requirements.txt
djangorestframework
Django~=2.2
psycopg2